Flytter K9Beskjed ut til dittNav mappe. Lager egen OmsorgspengerMidlertidigAleneBeskjedProperties
