+++ docker-compose-env/oneshot.json
@@ -83,6 +83,23 @@
         }
       ],
       "numPartitions": 3,
+      "topicName": "privat-omsorgspenger-midlertidig-alene-cleanup"
+    },
+    {
+      "configEntries": {
+        "retention.ms": -1
+      },
+      "members": [
+        {
+          "member": "igroup",
+          "role": "CONSUMER"
+        },
+        {
+          "member": "igroup",
+          "role": "PRODUCER"
+        }
+      ],
+      "numPartitions": 3,
       "topicName": "privat-k9-dittnav-varsel-beskjed"
     }
   ]
