+++ src/main/kotlin/no/nav/sifinnsynapi/dittnav/OmsorgspengerMidlertidigAleneBeskjedProperties.kt
@@ -6,10 +6,10 @@ import org.springframework.validation.annotation.Validated
 import java.net.URI
 
 @ConstructorBinding
-@ConfigurationProperties(prefix = "no.nav.dittnav.pleiepenger-sykt-barn.beskjed")
+@ConfigurationProperties(prefix = "no.nav.dittnav.omsorgspenger-midlertidig-alene.beskjed")
 @Validated
-data class PleiepengerDittnavBeskjedProperties(
+data class OmsorgspengerMidlertidigAleneBeskjedProperties (
         val tekst: String,
-        val link: URI,
-        val dagerSynlig: Long
+        val dagerSynlig: Long,
+        val link: URI? = null
 )
