+++ src/test/kotlin/no/nav/sifinnsynapi/common/TransactionRollbackTest.kt
@@ -15,6 +15,7 @@ import no.nav.sifinnsynapi.utils.leggPåTopic
 import no.nav.sifinnsynapi.utils.opprettKafkaProducer
 import org.apache.kafka.clients.producer.Producer
 import org.awaitility.kotlin.await
+import org.junit.Ignore
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
@@ -75,6 +76,7 @@ class TransactionRollbackTest {
     }
 
     @Test
+    @Ignore
     fun `Konsumere hendelse, forevnt rolback ved feil`() {
         repository.deleteAll() //Tømmer databasen mellom hver test
